- Separated css from their styled components and imported them into their corresponding components - Moved HomeScreen to pages, Added dummy Favorites page - Cleaned up timeSince helper - Centered gallery elements to view
