+++ src/hooks/useTheme.js
@@ -1,11 +1,13 @@
-import { useContext } from 'react';
+import { useContext, useState } from 'react';
 import { styles, themes } from '../constants/constants';
 import { EnvironmentContext } from '../contexts/EnvironmentContext';
 
 export const useTheme = () => {
   const { theme } = useContext(EnvironmentContext);
+
   const getTheme = () => {
     return theme === themes.light ? styles.colors.light : styles.colors.dark;
   };
-  return { getTheme };
+  const [cssTheme] = useState(getTheme());
+  return { cssTheme, getTheme };
 };
