+++ src/routers/AppRouter.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import { Router, Switch, Route, Redirect } from 'react-router-dom';
-import { AuthRouter } from './AuthRouter';
+import { Router, Switch, Route /*, Redirect */ } from 'react-router-dom';
+// import { AuthRouter } from './AuthRouter';
 import { createBrowserHistory } from 'history';
-import { HomeScreen } from '../components/HomeScreen';
+import { HomeScreen } from '../pages/HomeScreen';
+import { FavoritesScreen } from '../pages/FavoritesScreen';
 
 const history = createBrowserHistory();
 
@@ -11,9 +12,10 @@ export const AppRouter = () => {
     <Router history={history}>
       <div>
         <Switch>
-          <Route path="/auth" component={AuthRouter} />
+          {/* <Route path="/auth" component={AuthRouter} /> */}
+          <Route exact path="/favorites" component={FavoritesScreen} />
           <Route exact path="/" component={HomeScreen} />
-          <Redirect to="/auth/login" />
+          {/* <Redirect to="/auth/login" /> */}
         </Switch>
       </div>
     </Router>
