+++ src/styles/components/ui/gallery/Gallery.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const StyledGallery = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin-top: 45px;
+`;
