+++ src/styles/pages/Common.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const StyledContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export const StyledMain = styled.main`
+  flex: 1;
+  display: flex;
+  flex-direction: row;
+  margin-top: 45px;
+`;
